# lookup planning: adjust plan's intersection cost