--- a/pkg/ingester/lookupplan/mock_statistics_test.go
+++ b/pkg/ingester/lookupplan/mock_statistics_test.go
@@ -220,8 +220,8 @@ func newHighCardinalityMockStatistics() *mockStatistics {
 	}
 
 	stats.seriesPerValue["namespace"] = map[string]uint64{
-		"monitoring":   1500000,
-		"default":      1200000,
+		"monitoring":   1_500_000,
+		"default":      1_200_000,
 		"kube-system":  800000,
 		"logging":      600000,
 		"ingress":      400000,