… pkg/ingester/lookupplan/testdata/plan_cost_test_cases.csv
… pkg/ingester/lookupplan/testdata/planner_test_cases.csv